requirements.txt
requests
pandas
python-dotenv
tqdm
openai
ratelimit
pytest
